Testing LIDAR with PlatformIO
yesterday i tried to test the LIDAR VL53L4CD from the NUCLEO-F401RETX pack with the STM Cube IDE but it was way too complicated and i couldn't manage to even get the LIDAR to fire up, let alone display anything.
Today I switched to something a lot simpler because the point of this testing is to just give me hope that my parts work and manage to integrate them with each other, so I just settled for this simple but effective VSC extension.